% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friedman-nemenyi.R
\name{friedman.nemenyi}
\alias{friedman.nemenyi}
\title{Perform Friedman Test Followed by Nemenyi Post-Hoc Test}
\usage{
friedman.nemenyi(data, save)
}
\arguments{
\item{data}{A data frame containing the results of multiple methods. Each
row should represent a different method, and columns should represent the
observations or measurements.}

\item{save}{A string specifying the base name for output files. The function
will use this base name to save the difference matrix, Critical Difference
plot, and p-values plot.}
}
\value{
A data frame with the following columns:
  \itemize{
    \item \code{f.ChiSquare}: The Chi-Square statistic from the Friedman test.
    \item \code{f.pValue}: The p-value from the Friedman test.
    \item \code{f.Method}: The method used for the Friedman test.
    \item \code{n.CriticalDifference}: The Critical Difference from the Nemenyi test.
    \item \code{result}: A string indicating whether the methods are significantly different.
  }
}
\description{
This function performs the Friedman test on the provided data to assess
if there are significant differences among multiple methods. If the
Friedman test indicates significant differences, the function then performs
the Nemenyi post-hoc test to determine which methods are significantly different.
The results are saved as CSV and PDF files, including the difference matrix,
Critical Difference (CD) plot, and p-values plot.
}
\details{
The function saves three outputs:
\itemize{
  \item A CSV file with the name \code{paste(save, "-diff-matrix.txt", sep="")} containing
    the difference matrix from the Nemenyi post-hoc test.
  \item A PDF file with the name \code{paste(save, ".pdf", sep="")} containing the Critical
    Difference (CD) plot.
  \item A PDF file with the name \code{paste(save, "-pValues.pdf", sep="")} containing
    the p-values plot.
}
}
\examples{
# Example usage
# Assume 'my_data' is a data frame with results for different methods
# and 'base_name' is the base name for output files
result <- friedman.nemenyi(my_data, "base_name")

}
