% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{generate.rank.again}
\alias{generate.rank.again}
\title{Generate Rankings for CSV Files and Save Results}
\usage{
generate.rank.again(
  type,
  source.folder,
  destination.folder,
  file.names,
  names.list
)
}
\arguments{
\item{type}{A vector of column indices or names to select from the CSV files
for ranking. This determines which columns are included in the ranking.}

\item{source.folder}{The folder path where the input CSV files are located.}

\item{destination.folder}{The folder path where the ranking results will be saved.
Default is set to 'Folders', which should be defined in the user's environment.}

\item{file.names}{A vector of file names (without path) for the CSV files to be 
processed.}

\item{names.list}{A vector of names corresponding to the measures for each 
CSV file.}
}
\value{
A list containing ranking results for each file. The list has an 
  entry for each file, with the rankings stored as data frames.
}
\description{
This function processes a list of CSV files, generates rankings based on
specified columns, and saves the results in a destination folder. It handles
different types of ranking based on provided measure values.
}
\details{
The function reads CSV files from the `source.folder`, processes each file by
generating rankings based on the specified columns (`type`), and saves the 
rankings to the `destination.folder`. The measure values from `names.list` determine
which ranking is saved. If the measure value is 1, it saves the rankings in 
reverse order; otherwise, it saves them in the original order.
}
\examples{
# Example usage
type <- c("Column1", "Column2")
source.folder <- "path/to/source"
destination.folder <- "path/to/destination"
file.names <- c("file1.csv", "file2.csv")
names.list <- c("measure1", "measure2")

# Generate and save rankings
results <- generate.rank.again(type, source.folder, destination.folder, file.names, names.list)

}
