% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{generates.rank.again}
\alias{generates.rank.again}
\title{Generate Rankings for CSV Files and Save Results}
\usage{
generates.rank.again(
  type,
  source_folder,
  destination_folder = Folders,
  file_names,
  names_list
)
}
\arguments{
\item{type}{A vector of column indices or names to select from the CSV files
for ranking. This determines which columns are included in the ranking.}

\item{source_folder}{The folder path where the input CSV files are located.}

\item{destination_folder}{The folder path where the ranking results will be saved.
Default is set to 'Folders', which should be defined in the user's 
environment.}

\item{file_names}{A vector of file names (without path) for the CSV files to be 
processed.}

\item{names_list}{A vector of names corresponding to the measures for each 
CSV file.}
}
\value{
A list containing ranking results for each file. The list has an 
  entry for each file, with the rankings stored as data frames.
}
\description{
This function processes a list of CSV files, generates rankings based on
the provided columns, and saves the results in a specified destination folder.
It handles different types of ranking based on the provided measure values.
}
\details{
The function reads CSV files from the `source_folder`, processes each file by
generating rankings based on the specified columns (`type`), and saves the 
rankings to the `destination_folder`. The measure values from `names_list` determine
which ranking is saved. If the measure value is 1, it saves the rankings in 
reverse order; otherwise, it saves them in the original order.
}
\examples{
# Example usage
type <- c("Column1", "Column2")
source_folder <- "path/to/source"
destination_folder <- "path/to/destination"
file_names <- c("file1.csv", "file2.csv")
names_list <- c("measure1", "measure2")

# Generate and save rankings
results <- generates.rank.again(type, source_folder, destination_folder, file_names, names_list)

}
